fix(categories): fix 0% overview display after editing a category - Remove expenseOnThisCategory property from Category class - Add Long id property to Transaction and Category classes, remove TransactionInput class - Add extension function to find and verify category existence by id

+++ app/src/main/java/com/example/task_1/ui/category/CategoryDeleteDialog.kt
@@ -12,10 +12,10 @@ import com.example.task_1.domain.Category
 // TODO make it usable for transactions as well
 @Composable
 fun CategoryDeleteDialog(
-    categoryIDForDeletion: Int,
+    categoryIDForDeletion: Long,
     currentCategory: Category,
     closeDialog: () -> Unit,
-    removeCategory: (Int) -> Unit
+    removeCategory: (Long) -> Unit
 ) {
 
     AlertDialog(
